Reorganize preprocessors to folders
