+++ src/lib/preprocessors/texts/lemmatization.py
@@ -5,9 +5,9 @@
 #
 
 
-from .preprocessor import Preprocessor
-from ..utils.iterator_utils import apply
-from ..utils.spacy_utils import get_spacy_nlp
+from ..preprocessor import Preprocessor
+from ...utils.iterator_utils import apply
+from ...utils.spacy_utils import get_spacy_nlp
 
 
 class Lemmatization(Preprocessor):
