+++ src/lib/var/preprocessors.json
@@ -1,13 +1,13 @@
 {
     "divide": "lib.preprocessors.divide.Divide",
-    "grayscale": "lib.preprocessors.grayscale.Grayscale",
-    "image_resize": "lib.preprocessors.image_resize.ImageResize",
+    "grayscale": "lib.preprocessors.images.grayscale.Grayscale",
+    "image_resize": "lib.preprocessors.images.image_resize.ImageResize",
     "normalize": "lib.preprocessors.normalize.Normalize",
     "replace": "lib.preprocessors.replace.Replace",
-    "grayscale_to_color": "lib.preprocessors.grayscale_to_color.GrayscaleToColor",
-    "lemmatization": "lib.preprocessors.lemmatization.Lemmatization",
-    "alphanumeric": "lib.preprocessors.alphanumeric.Alphanumeric",
-    "tfidf": "lib.preprocessors.tfidf.Tfidf",
+    "grayscale_to_color": "lib.preprocessors.images.grayscale_to_color.GrayscaleToColor",
+    "lemmatization": "lib.preprocessors.texts.lemmatization.Lemmatization",
+    "alphanumeric": "lib.preprocessors.texts.alphanumeric.Alphanumeric",
+    "tfidf": "lib.preprocessors.texts.tfidf.Tfidf",
     "kbest": "lib.preprocessors.kbest.KBest",
     "standard_scaler": "sklearn.preprocessing.StandardScaler"
 }
